@@ -5,7 +5,7 @@ author:
 homepage:
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.6.0 <3.0.0"
 
 flutter_intl:
   enabled: true
@@ -30,9 +30,12 @@ dependencies:
   # https://github.com/dart-lang/convert
   convert: ">=1.0.0 <4.0.0"
 
+  dartx: ^0.4.1
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  pedantic: ^1.8.0  
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
