@@ -19,11 +19,12 @@ typedef String MessageIfAbsent(String messageStr, List<dynamic> args);
 class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'en';
 
-  final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
+  final messages = _notInlinedMessages();
+  static Map<String, Function> _notInlinedMessages() => <String, Function> {
     "copied" : MessageLookupByLibrary.simpleMessage("copied"),
     "copy" : MessageLookupByLibrary.simpleMessage("copy"),
     "showLess" : MessageLookupByLibrary.simpleMessage("show less"),
-    "showMore" : MessageLookupByLibrary.simpleMessage("show more")
+    "showMore" : MessageLookupByLibrary.simpleMessage("show more"),
+    "test" : MessageLookupByLibrary.simpleMessage("test")
   };
 }
